fixed datatypes.java bugs

+++ src/DataTypes.java
@@ -1,6 +1,9 @@
+import org.junit.jupiter.api.Test;
+import static org.junit.jupiter.Assertions.asserEquals;
+
 public class DataTypes {
     public static long sum(int[] numbers) {
-        long s = 0; // variable to accumulate the sum in!
+        int s = 0; // variable to accumulate the sum in!
         // below is a "foreach" loop which iterates through numbers
         for (int x : numbers) {
             s += x;
